+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # v1.0.3:
 - It is no longer possible to register with a character that is already registered
+- Team captains can now select a map
+- Added =confirm command for Team Captains to agree on a map
+- Added role updates when agreeing with the rules
+- Added notify feature
 
 # v1.0.2:
 - Now properly checking if player have no missing faction when registering with a Jaeger char